+++ src/demo/src/examples/Success.tsx
@@ -3,7 +3,7 @@ import {Example} from "./Example";
 const title: string = "A success message!";
 
 const snippet: string = `
-<SweetAlert success title="Good job!" onConfirm={this.onConfirm}>
+<SweetAlert success title="Good job!" onConfirm={this.onConfirm} onCancel={this.onCancel}>
   You clicked the button!
 </SweetAlert>
 `;